requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.2.10
asttokens==2.0.5
async-timeout==4.0.2
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
blis==0.7.9
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
comm==0.1.3
confection==0.0.4
cryptography==41.0.1
cymem==2.0.7
dataclasses-json==0.5.14
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
Django==3.1
dnspython==2.4.1
docx==0.2.4
docx2txt==0.8
email-validator==2.0.0.post2
entrypoints==0.4
exceptiongroup==1.1.2
executing==0.8.3
faiss-cpu==1.7.2
fastapi==0.101.0
fastjsonschema==2.17.1
filelock==3.9.0
filetype==1.2.0
fqdn==1.5.1
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
ipykernel==6.23.1
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.3.1
jsonpointer==2.3
jsonschema==4.18.6
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.6.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
langchain==0.0.253
langcodes==3.3.0
langsmith==0.0.19
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.3
mdurl==0.1.2
mistune==2.0.5
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.4.0
nbformat==5.9.0
nest-asyncio==1.5.5
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.25.2
openai==0.27.8
openapi-schema-pydantic==1.2.4
orjson==3.9.3
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathy==0.10.1
pdfminer.six==20221105
pickleshare==0.7.5
Pillow==10.0.0
pipenv==2023.2.4
platformdirs==2.6.2
pluggy==1.0.0
preshed==3.0.8
prometheus-client==0.17.0
prompt-toolkit==3.0.38
protobuf==4.23.4
psutil==5.9.0
pure-eval==0.2.2
pyarrow==12.0.1
pycparser==2.21
pycryptodome==3.18.0
pydantic==2.1.1
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.4.0
pydeck==0.8.0
pygame==2.1.2
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.8
pypdf==3.15.0
PyPDF2==3.0.1
pyresparser==1.0.6
pyrsistent==0.19.3
pytest==7.3.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pywinpty==2.0.10
PyYAML==6.0.1
pyzmq==25.1.0
qtconsole==5.4.3
QtPy==2.3.1
referencing==0.30.2
regex==2023.6.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.9.2
Send2Trash==1.8.2
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.19
sqlparse==0.3.1
srsly==2.4.6
stack-data==0.2.0
starlette==0.27.0
tabulate==0.9.0
tenacity==8.2.2
terminado==0.17.1
thinc==8.1.10
tiktoken==0.4.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
typer==0.7.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
ujson==5.8.0
unstructured==0.9.0
uri-template==1.2.0
urllib3==2.0.4
uvicorn==0.23.2
validators==0.20.0
virtualenv==20.17.1
virtualenv-clone==0.5.7
wasabi==1.1.2
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.2.5
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.2
websockets==11.0.3
widgetsnbextension==4.0.7
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.16.2
